chore: consolidate duplicate code (#907)
* chore: consolidate duplicated code

* chore: fix data

* chore: sort imports

* chore: imports

* chore: group imports
