+++ types/ormstore/db.go
@@ -2,15 +2,13 @@ package ormstore
 
 import (
 	"context"
+
 	ormv1alpha1 "github.com/cosmos/cosmos-sdk/api/cosmos/orm/v1alpha1"
-	"github.com/regen-network/regen-ledger/types"
-
-	queryv1beta1 "github.com/cosmos/cosmos-sdk/api/cosmos/base/query/v1beta1"
-	"github.com/cosmos/cosmos-sdk/types/query"
-
 	"github.com/cosmos/cosmos-sdk/orm/model/ormdb"
 	"github.com/cosmos/cosmos-sdk/orm/model/ormtable"
 	storetypes "github.com/cosmos/cosmos-sdk/store/types"
+
+	"github.com/regen-network/regen-ledger/types"
 )
 
 // NewStoreKeyDB creates an ormdb.ModuleDB from an ormdb.ModuleDB and a StoreKey.
@@ -31,21 +29,4 @@ func NewStoreKeyDB(desc *ormv1alpha1.ModuleSchemaDescriptor, key storetypes.Stor
 	}
 	options.GetBackendResolver = backEndResolver
 	return ormdb.NewModuleDB(desc, options)
-}
-
-func GogoPageReqToPulsarPageReq(request *query.PageRequest) *queryv1beta1.PageRequest {
-	return &queryv1beta1.PageRequest{
-		Key:        request.Key,
-		Offset:     request.Offset,
-		Limit:      request.Limit,
-		CountTotal: request.CountTotal,
-		Reverse:    request.Reverse,
-	}
-}
-
-func PulsarPageResToGogoPageRes(response *queryv1beta1.PageResponse) *query.PageResponse {
-	return &query.PageResponse{
-		NextKey: response.NextKey,
-		Total:   response.Total,
-	}
 }
