+++ types/ormutil/compatability.go
@@ -1,11 +1,10 @@
-package core
+package ormutil
 
 import (
-	gogoproto "github.com/gogo/protobuf/proto"
-	"google.golang.org/protobuf/proto"
-
 	queryv1beta1 "github.com/cosmos/cosmos-sdk/api/cosmos/base/query/v1beta1"
 	"github.com/cosmos/cosmos-sdk/types/query"
+	gogoproto "github.com/gogo/protobuf/proto"
+	"google.golang.org/protobuf/proto"
 )
 
 func GogoPageReqToPulsarPageReq(from *query.PageRequest) (*queryv1beta1.PageRequest, error) {
