+++ x/ecocredit/server/core/query_batches.go
@@ -2,15 +2,17 @@ package core
 
 import (
 	"context"
+
 	"github.com/cosmos/cosmos-sdk/orm/model/ormlist"
 
 	api "github.com/regen-network/regen-ledger/api/regen/ecocredit/v1"
+	"github.com/regen-network/regen-ledger/types/ormutil"
 	"github.com/regen-network/regen-ledger/x/ecocredit/core"
 )
 
 // Batches queries for all batches in the given credit class.
 func (k Keeper) Batches(ctx context.Context, request *core.QueryBatchesRequest) (*core.QueryBatchesResponse, error) {
-	pg, err := GogoPageReqToPulsarPageReq(request.Pagination)
+	pg, err := ormutil.GogoPageReqToPulsarPageReq(request.Pagination)
 	if err != nil {
 		return nil, err
 	}
@@ -31,12 +33,12 @@ func (k Keeper) Batches(ctx context.Context, request *core.QueryBatchesRequest) 
 		}
 
 		var bi core.BatchInfo
-		if err = PulsarToGogoSlow(batch, &bi); err != nil {
+		if err = ormutil.PulsarToGogoSlow(batch, &bi); err != nil {
 			return nil, err
 		}
 		batches = append(batches, &bi)
 	}
-	pr, err := PulsarPageResToGogoPageRes(it.PageResponse())
+	pr, err := ormutil.PulsarPageResToGogoPageRes(it.PageResponse())
 	if err != nil {
 		return nil, err
 	}
