+++ x/ecocredit/server/core/query_batches_by_class.go
@@ -2,14 +2,17 @@ package core
 
 import (
 	"context"
+
 	"github.com/cosmos/cosmos-sdk/orm/model/ormlist"
+
 	api "github.com/regen-network/regen-ledger/api/regen/ecocredit/v1"
+	"github.com/regen-network/regen-ledger/types/ormutil"
 	"github.com/regen-network/regen-ledger/x/ecocredit/core"
 )
 
 // BatchesByClass queries all batches issued under a given credit class.
 func (k Keeper) BatchesByClass(ctx context.Context, request *core.QueryBatchesByClassRequest) (*core.QueryBatchesByClassResponse, error) {
-	pg, err := GogoPageReqToPulsarPageReq(request.Pagination)
+	pg, err := ormutil.GogoPageReqToPulsarPageReq(request.Pagination)
 	if err != nil {
 		return nil, err
 	}
@@ -30,12 +33,12 @@ func (k Keeper) BatchesByClass(ctx context.Context, request *core.QueryBatchesBy
 		}
 
 		var bi core.BatchInfo
-		if err = PulsarToGogoSlow(batch, &bi); err != nil {
+		if err = ormutil.PulsarToGogoSlow(batch, &bi); err != nil {
 			return nil, err
 		}
 		batches = append(batches, &bi)
 	}
-	pr, err := PulsarPageResToGogoPageRes(it.PageResponse())
+	pr, err := ormutil.PulsarPageResToGogoPageRes(it.PageResponse())
 	if err != nil {
 		return nil, err
 	}
