+++ x/ecocredit/server/core/query_class_info.go
@@ -2,6 +2,8 @@ package core
 
 import (
 	"context"
+
+	"github.com/regen-network/regen-ledger/types/ormutil"
 	"github.com/regen-network/regen-ledger/x/ecocredit/core"
 )
 
@@ -13,7 +15,7 @@ func (k Keeper) ClassInfo(ctx context.Context, request *core.QueryClassInfoReque
 	}
 
 	var ci core.ClassInfo
-	if err = PulsarToGogoSlow(classInfo, &ci); err != nil {
+	if err = ormutil.PulsarToGogoSlow(classInfo, &ci); err != nil {
 		return nil, err
 	}
 	return &core.QueryClassInfoResponse{Info: &ci}, nil
